requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3
anthropic==0.34.2
openai>=1.30.0
python-dotenv==1.0.1
